Comment out messages index socket and put one in top nav
